requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Geospatial libraries
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
osmnx>=1.7.0
folium>=0.15.0

# Graph and routing
networkx>=3.2.0
ortools>=9.8.0

# Data processing
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0
scikit-learn>=1.3.0

# HTTP requests and utilities
requests>=2.31.0
loguru>=0.7.0
python-dotenv>=1.0.0
PyJWT>=2.8.0

# Database (optional)
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0

# Additional utilities
polyline>=2.0.0